Refactor custom ops classes to use python_op_factory as base (#5338)
`python_op_factory` was extended and documented. There are new, optional arguments:
* `internal_schema_name` that will be used to retrieve schema and spec for argument
   handling on Python side, while allowing the original `schema` to be used for the purpose
   of exposing the documentation.
* `generated` - to mark that this is fully automatic binding or it is extended by hand.

As both `TFRecordReader` and `PythonFunction` has several variants with different base
implementation schemas, the base classes are converted into a class generator function. 
The new classes are marked as not generated, as the type hints are done manually for them.

This reduces the custom code to minimum, allowing the base class to handle all properties
and some common arguments. The argument handling is moved to base class by updating
the kwargs before invoking the base class.

When a `tfrec.Feature` is encountered in the Python argument serialization layer, it is again
processed by the `tfrec.Feature` constructor (as we do for other types, normalizing numbers
with int() or float()). For this purpose a copy constructor is added and exposed in Pybind.
An alternative would be to change the conversion to an identity function - that would keep
the error in the spec.AddArg rather than moving it to constructor parameter matching.
Validation code is added to the operator, to check for matching type with better error message.

NumbaFunction was also reworked, inheriting directly from the python_op_factory.

Calls to `_raise_no_current_pipeline` were eliminated - this function doesn't exist!

Breaking change: There is no longer a `PythonFunctionBase` class in `nvidia.dali.ops`.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>