Tasking module (#5436)
# Tasking module
`dali::tasking` is a task execution engine which supports inter-task dependencies, semaphores and data passing.
Main concepts of the tasking API are:
* Task - a single-use object that encapsulates a function; it can depend on zero or more Waitable objects
* Waitable - an object for which a task can wait
* Scheduler - manages tasks and resolves their readiness state

## Main components
A `Task` is also `Waitable` - that is, a task can depend on the completion of other task(s).
There are two kinds of `Waitable` objects - completion events and releasable objects.
A `CompletionEvent` (e.g. Task) is an object which remains acquirable after it's "completed". Performing Acquire operation on it does not alter the state of the object.
A `Semaphore` can be externally released (it's `Releasable`), but `Acquire` lowers the semaphore count, so the number of tasks that can acquire a semaphore before it's released is limited.

Acquisition of `Waitable` objects on behalf of waiting tasks is the duty of the `Scheduler`. By contrast, `Release` operation on `Releasable` objects can be performed in any context.

## Data passing
Data can be passed between tasks. It's wrapped in a `TaskResult` type which stores either the value as `std::any` or the error as `std::exception_ptr`. An attempt to access the value with an exception present results in the exception being rethrown.
A `Task` wraps a function taking 0 or 1 argument (in the latter case it's this-like pointer to the `Task`) and returning a value of values.
Upon creation of the task, the caller defines the number of return values. There are two options:
- Scalar return value - in which case the return value can be any objects convertible to `std::any`. If the task returns `std::any`, it's stripped and the stored value is copied.
- Multiple return values - in this case the return value of the function must be either a collection or a `tuple`.
A `Task` can `Subscribe` to the result of a producer task (which has to be done before the producer is submitted to the scheduler). The result can be obtained by calling `task->GetInputValue<T>(index)`

## Obtaining results
When a task is added to the scheduler, a `TaskFuture` object can be obtained. `Scheduler::AddSilentTask` should be used if the result of the task is not needed. Calling `TaskFuture::Value` will wait for the task to complete and return the value produced by the task (or rethrow the exception).

## Error handling
All errors thrown by the task functions are stored in the task results. An attempt to obtain results of a task that threw an exception will result in the exception being rethrown.

---------

Signed-off-by: Michal Zientkiewicz <michalz@nvidia.com>