Enable runtime and sphinx-level signatures for ops API
This should allow proper rendering of the parameter references in the
ops API part of the documentation. Additionally this fixes lack of
default values for some CMAKE build flags.

If the `BUILD_WITH_ASAN` doesn't have a value
`if(NOT ${CMAKE_CROSSCOMPILING} AND NOT ${BUILD_WITH_ASAN})` does
not evaluate to true and dali/python/CMakeLists.txt skips the generation
of interface files with API signatures - breaking local developement.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>