Use absolute addressing for parameters
Injecting full qualified path to parameters solve problems with lookup
of enclosing entity and generating valid links for modules that are
runtime generated.

The missing-reference resolver of sphinx paramlinks assume that `source`
of all nodes has format:
"/path/to/file.py:docstring of module.clsname.methname"
and requires the `:`, while members of generated DALI submodules have
only the:
"docstring of nvidia.dali.fn.decoders.image"
part, so the regex doesn't match. If the reference is fully specified,
the plugin doesn't need to look it up and doesn't use the unsuitable
regex.

Some nvidia.dali.ops module specific code is added, so we try to always
obtain the `__call__` signature (it contains all inputs and kwargs) and
we try to link to either class or `__call__` documentation depending if
we are documenting input or kwarg - as that API has this specific
distinction.

Other parts of API don't have specific conventions.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>