CSE - Common Subexpression (Subgraph) Elimination (#5752)
This change introduces a graph-level optimization that removes common subgraphs.
The algorithm goes over graph nodes in topological order and computes a key from the operator's schema name, arguments and (renamed) inputs. If an existing operator is found, the current one is replaced with the previously found one and the output names are added to the renaming map. When subsequent operators are processed, the inputs of the replaced operator are renamed accordingly.
Example:

A = foo(arg=42)  # new
B = foo(arg=42)  # replaced with A
C = bar(A) # new
D = bar(B) # replaced with bar(A)
return C + D  # replaced with C + C

This optimization is particularly useful with constants, the use of .shape function and subscripts.
In the following example:

horz_center = (bbox[0] + bbox[2]) / 2
width = (bbox[2]  - bbox[0])

The bbox[0] and bbox[2] will be reused.

---------

Signed-off-by: Michal Zientkiewicz <michalz@nvidia.com>