Fix data race in experimental image decoder (#5856)
Fixes a rare data corruption issue observed in experimental image decoder.
Added a WaitForWork at the end of RunImpl in image decoder, which makes sure that nvimagecodec is done using our thread pool executor.
nvimagecodec currently has a bug that sets the promise for the batch before it has issued stream synchronization with the user stream. Adding WaitForWork forces us to wait for it to finish.

Signed-off-by: Joaquin Anton Guirao <janton@nvidia.com>