Imperative mode: Invocation (#6042)
This commit adds the Invocation class, which binds an operator instance to a set of run-time arguments.
It supports lazy evaluation and provides access to individual results of multivalent operators.
The InvocationResult class describes a single result - when the result is requested, the invocation is executed. Multiple invocation results pointing to a single invocation avoid multiple evaluation.
The tests use mock types for Tensor, Batch and Operator.
-----
Signed-off-by: Michal Zientkiewicz <michalz@nvidia.com>

Signed-off-by: Tony Reina <tony.reina@exthymic.com>