Dynamic Mode: Tensor and Batch Types (#6056)
This PR brings two classes that are foundational for the Dynamic Mode: Tensor and Batch.
Both classes can wrap either an actual object, backed by Tensor[List]<CPU|GPU> or InvocationResult or results of indexing or composition.
A batch can be:

* TensorList<CPU|GPU>
* a list of Tensor objects

A tensor can be:
* Tensor<CPU|GPU>
* a Batch and an index in the batch
* a TensorSlice object

There are also utility functions, tensor, as_tensor and batch, as_batch. The ones without as_ construct a tensor or batch, respectively, and perform a copy. The functions with as_ prefix avoid the copy, if possible.

Batch access:
as of this PR, Batch deliberately doesn't include __getitem__ or __len__. It can be iterated to yield tensors, but it cannot be indexed. To access the list of tensors, use .tensors property; to apply batched slicing, use .slice property. __getitem__ will be added later and will implement either tensor access or batched slicing, which remains to be decided.

------
Signed-off-by: Michal Zientkiewicz <michalz@nvidia.com>