fn.bbox_rotate (#5979)
Add bounding box rotation as a single op with options including how they should expand with the rotate, whether the image canvas was fixed when the image was rotated, and the box format (layout and normalization).

This op prunes boxes and labels if keep_size=True and the box is truncated to a fraction below remove_threshold.
---------

Signed-off-by: Bryce Ferenczi <frenzi@hotmail.com.au>