Random operator rework (#6100)
Random number generators now take a `_rng_state` argument, which, when provided, overrides the seed and makes the upcoming iteration totally deterministic and repeatable.
To achieve that, the generators use a counter-based Philox RNG and the random state consists of the key and counter.

This PR reworks the RNGBase, which no longer needs to maintain a state array (as states can be simply regenerated each time).

-----
Signed-off-by: Michal Zientkiewicz <michalz@nvidia.com>