Fix enum handling in Dynamic Mode + require NumPy (#6150)
This PR fixes the handling of DALI enums in NDD tensors and batches.
Prior to this change, when dtype was passed to Tensor constructor, DALI enums were not properly converted and an error was raised. This PR fixes that and adds tests for Tensor and Batch.

This PR also adds NumPy as an explicit DALI dependency (we require it anyway to perform basic tasks).

Signed-off-by: Michał Zientkiewicz <mzient@gmail.com>